requirements.txt
azure-mgmt-network
azure-identity
azure-mgmt-resource
